requirements.txt
adbutils==0.11.0
logzero==1.7.0
APScheduler==3.8.0
baidu-aip==2.2.18.0
numpy==1.19.3
Flask==2.0.2
opencv-python==4.5.3.56
Pillow==8.3.2
retrying==1.3.3
SQLAlchemy==1.4.25
Flask-Babel==2.0.0
Flask-SocketIO==5.0.1
PyYAML==5.4.1
yagmail==0.14.260
pytz~=2021.3
eventlet==0.32.0
cnocr==2.1.0
cnstd==1.1.1
flask-cors==3.0.10
psutil==5.9.0